layout for android #2

+++ Mark5.Mobile.IOS/Ui/ViewControllers/DownloadViewController.cs
@@ -562,7 +562,7 @@ namespace Mark5.Mobile.IOS.Ui.ViewControllers
 
                         progressLabel.Text = Localization.GetString("downloading_percentage___", (int)(progressIndicator.Progress * 100));
                         if (timeLeft > 0)
-                            progressLabel.Text += " " + Localization.GetString("time_remaining_minutes", timeLeft);
+                            progressLabel.Text += " " + Localization.GetString("time_remaining_minutes", (int)timeLeft);
                         else if (timeLeft > -1)
                             progressLabel.Text += " " + Localization.GetString("time_remaining_less_than_minute");
                     }
